% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pipe_between}
\alias{.pipe_between}
\title{Quote a chain of (abritrary number of) pipes.}
\usage{
.pipe_between(x, ...)
}
\arguments{
\item{x}{Starting function for the pipe chain.}

\item{...}{Subsequent functions appended to the pipe chain.}
}
\value{
A quoted pipe chain starting with \code{x} to the last element of \code{...},
or \code{x} itself if \code{...} is empty.
}
\description{
Quote a chain of (abritrary number of) pipes.
}
\examples{
\dontrun{
# Creating pipe chain
.pipe_between(2, runif())
#> 2 \%>\% runif()

# Evaluating pipe chain later on
eval(.pipe_between(2, runif()))
#> [1] 0.5180003 0.5353272

# Longer pipe chain
long_chain <- .pipe_between(2, runif(), sum(), log())
long_chain
#> 2 \%>\% runif() \%>\% sum() \%>\% log()

eval(long_chain)
#> [1] -0.4767814
}

}
