% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_augment.R
\name{tar_augment}
\alias{tar_augment}
\title{Predict Target}
\usage{
tar_augment(name, x, new_data, ..., tar_args = list())
}
\arguments{
\item{name}{Name of target}

\item{x}{A \code{model_fit} object produced by \code{\link[parsnip:reexports]{fit()}} or \code{\link[parsnip:reexports]{fit_xy()}}.}

\item{new_data}{A data frame or matrix.}

\item{...}{Not currently used.}

\item{tar_args}{Arguments passed to \code{tar_target}.}
}
\description{
Predict Target
}
