% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_fit.R
\name{tar_fit_xy}
\alias{tar_fit_xy}
\title{Fitting XY Target}
\usage{
tar_fit_xy(
  object,
  x,
  y,
  control = parsnip::control_parsnip(),
  ...,
  tar_args = list()
)
}
\arguments{
\item{object}{An object of class \code{model_spec} that has a chosen engine
(via \code{\link[parsnip:set_engine]{set_engine()}}).}

\item{x}{A matrix, sparse matrix, or data frame of predictors. Only some
models have support for sparse matrix input. See \code{parsnip::get_encoding()}
for details. \code{x} should have column names.}

\item{y}{A vector, matrix or data frame of outcome data.}

\item{control}{A named list with elements \code{verbosity} and
\code{catch}. See \code{\link[parsnip:control_parsnip]{control_parsnip()}}.}

\item{...}{Not currently used; values passed here will be
ignored. Other options required to fit the model should be
passed using \code{set_engine()}.}

\item{tar_args}{Arguments passed to \code{tar_target}.}
}
\value{
A \code{tar_target} object.
}
\description{
Fitting XY Target
}
