% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_model.R
\name{tar_model}
\alias{tar_model}
\title{Model Target}
\usage{
tar_model(name, model, ..., tar_args = list())
}
\arguments{
\item{name}{Name of target}

\item{model}{Parsnip model function.}

\item{...}{Additional functions that \code{model} is passed to.}

\item{tar_args}{Arguments passed to \code{tar_target}.}
}
\value{
A \code{tar_target} object.
}
\description{
Model Target
}
\examples{
\dontrun{
    # Create a ranger model that is saved to a .qs
    tar_model(
        make_model_rf,
        rand_forest,
        set_mode("regression"),
        set_engine("ranger"),
        tar_args = list(format = "qs")
    )

    # Create a nnet neural network model
    tar_model(
        make_model_mlp,
        mlp,
        set_args(hidden_units = 5),
        set_engine("nnet"),
        set_mode("regression")
    )
}
}
