% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_predict.R
\name{tar_predict}
\alias{tar_predict}
\title{Predict Target}
\usage{
tar_predict(
  name,
  object,
  new_data,
  type = NULL,
  opts = list(),
  ...,
  tar_args = list()
)
}
\arguments{
\item{name}{Name of target}

\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{type}{A single character value or \code{NULL}. Possible values
are "numeric", "class", "prob", "conf_int", "pred_int", "quantile",
or "raw". When \code{NULL}, \code{predict()} will choose an appropriate value
based on the model's mode.}

\item{opts}{A list of optional arguments to the underlying
predict function that will be used when \code{type = "raw"}. The
list should not include options for the model object or the
new data being predicted.}

\item{...}{
  Arguments passed on to \code{\link[parsnip:predict.model_fit]{parsnip::predict.model_fit}}
  \describe{
    \item{\code{}}{}
  }}

\item{tar_args}{Arguments passed to \code{tar_target}.}
}
\description{
Predict Target
}
