% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_recipe.R
\name{tar_recipe}
\alias{tar_recipe}
\alias{tar_recipe.default}
\alias{tar_recipe.formula}
\alias{tar_recipe.data.frame}
\alias{tar_recipe.matrix}
\title{Recipe Target}
\usage{
tar_recipe(name, x, ...)

\method{tar_recipe}{default}(name, x, ...)

\method{tar_recipe}{data.frame}(name, x, formula = NULL, ..., tar_args = list())

\method{tar_recipe}{formula}(name, formula, data, ..., tar_args = list())

\method{tar_recipe}{matrix}(name, x, ..., tar_args = list())
}
\arguments{
\item{name}{Name of target}

\item{x}{A data frame or tibble of the \emph{template} data set
(see below).}

\item{...}{Additional functions that the recipe is passed to,
i.e. recipe step functions.}

\item{formula}{A model formula. No in-line functions should be used here
(e.g. \code{log(x)}, \code{x:y}, etc.) and minus signs are not allowed. These types of
transformations should be enacted using \code{step} functions in this package.
Dots are allowed as are simple multivariate outcome terms (i.e. no need for
\code{cbind}; see Examples). A model formula may not be the best choice for
high-dimensional data with many columns, because of problems with memory.}

\item{tar_args}{Arguments passed to \code{tar_target}.}

\item{data}{A data frame or tibble of the \emph{template} data set
(see below).}
}
\value{
A \code{tar_target} object.
}
\description{
Recipe Target
}
\examples{
\dontrun{
    # Basic recipe
    tar_recipe(
        make_recipe,
        x = mtcars,
        formula = mpg ~ .,
        step_dummy(gear, carb),
        tar_args = list(format = "qs")
    )
}
}
