% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_workflow.R
\name{tar_workflow}
\alias{tar_workflow}
\title{Recipe Target}
\usage{
tar_workflow(name, ..., tar_args = list())
}
\arguments{
\item{name}{Name of target}

\item{...}{Functions piped to from \code{workflows::workflow()}
such as \code{workflows::add_model()}. See example.}

\item{tar_args}{Arguments passed to \code{tar_target}.}
}
\value{
A \code{tar_target} object.
}
\description{
Recipe Target
}
\examples{
\dontrun{
list(
    # `x` is some template data frame
    tar_recipe(
        make_recipe,
        V1 ~ V2 + V3,
        data = x
    ),
    # Creating a model
    tar_model(
        make_model_rf,
        rand_forest,
        set_mode("regression"),
        set_engine("ranger"),
        tar_args = list(format = "qs")
    ),
    # Creating the workflow
    tar_workflow(
        make_workflow_rf,
        workflows::add_model(make_model_rf),
        workflows::add_recipe(make_recipe)
    )
)
}
}
